Update fortuna.js
When reseeding the key, it would be good to retain some of the entropy in the current key. If fortuna.entropy has poor quality entropy for any reason, the security of the RNG will not be compromised. The entropy of the key would be at least as good as the previous key, with the addition of the entropy in fortuna.entropy.